Added a button to change current fractal.

+++ drawable2d.hpp
@@ -8,6 +8,6 @@ struct CanvasContext{
 };
 class Drawable2D{
     public:
-    virtual sf::Color get_pixel(const CanvasContext& ctx,double x,double y) const = 0;
+    virtual sf::Color get_pixel(const CanvasContext& ctx,int x,int y) const = 0;
     virtual ~Drawable2D(){};
 };
