+++ fractal.hpp
@@ -2,6 +2,6 @@
 #include "drawable2d.hpp"
 
 class Fractal:public Drawable2D{
-    virtual sf::Color get_pixel(const CanvasContext& ctx,double x,double y) const override = 0;
+    virtual sf::Color get_pixel(const CanvasContext& ctx,int x,int y) const override = 0;
 
 };
