+++ julia_fractal.hpp
@@ -6,6 +6,6 @@ class JuliaFractal:public Fractal{
     const unsigned int maxiter =1e2;
     public:
     JuliaFractal(double a=0,double b =0);
-    virtual sf::Color get_pixel(const CanvasContext& ctx,double x,double y) const override;
+    virtual sf::Color get_pixel(const CanvasContext& ctx,int x,int y) const override;
 
 };
