+++ mandelbrot_fractal.hpp
@@ -3,5 +3,5 @@
 class MandelbrotFractal:public Fractal{
     const int maxiter = 1e2;
     public:
-    virtual sf::Color get_pixel(const CanvasContext& ctx,double x,double y) const override;
+    virtual sf::Color get_pixel(const CanvasContext& ctx,int x,int y) const override;
 };
